+++ server/extensions.py
@@ -4,4 +4,4 @@ from flask_argon2 import Argon2
 
 argon2 = Argon2()
 db = SQLAlchemy()
-loginmgr = LoginManager()
+login_manager = LoginManager()
