+++ server/public/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+"""The public module, including the homepage and user auth."""
+from . import views  # noqa
