+++ server/server.py
@@ -2,12 +2,10 @@ from flask import Flask, request, send_from_directory
 from config import Config
 from flask_migrate import Migrate
 from flask_cors import CORS
-from flask_login import LoginManager, login_user, current_user,logout_user
 import os
 from src.dashboard.dashapp import create_dash_app
-from extensions import db, loginmgr, argon2
-from models import User
-
+from extensions import db, login_manager, argon2
+from server import user, public
 
 def create_app(config_object = Config):
 
@@ -15,66 +13,29 @@ def create_app(config_object = Config):
                 instance_relative_config=True)
     app.add_url_rule('/', 'root', lambda: app.send_static_file('index.html'))
     app.config.from_object(Config)
-    CORS(app)
-    app.route('/', defaults={'path': ''})
     register_extensions(app)
-    @app.route('/<path:path>')
-    def serve(path):
-        if path != "" and os.path.exists("src/client/" + path):
-            return send_from_directory('src/client', path)
-        else:
-            return send_from_directory('src/client', 'index.html')
-
-    @app.route('/signup', methods=['POST', 'GET'])
-    def signupfunc():
-        if request.method == 'POST':
-            robj = request.get_json()
-            user = User(username=robj['name'], email=robj['email'])
-            user.set_password(robj['password'])
-            # max_id = db.session.query(db.func.max(User.id)).scalar()
-            # user.set_user_id(max_id)
-            db.session.add(user)
-            db.session.commit()
-            print('signedup')
-
-        return 'signedup'
-
-    @app.route('/login', methods=['POST'])
-    def login():
-        # if current_user.is_authenticated:
-        #     print('alreadyauth')
-        #     return 'alreadyauth'
-        jobj = request.get_json()
-        user = User.query.filter_by(email=jobj['email']).first()
-        if user is None or not user.check_password(jobj['password']):
-            print("error")
-            return "Error"
-        else:
-            login_user(user)
-            print('loggedin')
-            return 'loggedin'
-
-    @app.route('/logout')
-    def logout():
-        logout_user()
-        return redirect(url_for('index'))
-
+    register_blueprints(app)
     return app
 
 
 def register_extensions(app):
     argon2.init_app(app)
-    loginmgr.init_app(app)
-    loginmgr.login_view = 'login'
+    login_manager.init_app(app)
+    login_manager.login_view = 'login'
     create_dash_app(app)
-
+    CORS(app)
     # Database initialisation
     db.init_app(app)
     migrate = Migrate(app, db)
     with app.app_context():
         db.create_all()
-    return db
+    return None
 
+def register_blueprints(app):
+    """Register Flask blueprints."""
+    app.register_blueprint(public.views.blueprint)
+    app.register_blueprint(user.views.blueprint)
+    return None
 
 if __name__ == '__main__':
     app = create_app(Config)
