+++ server/user/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+"""The user module."""
+from . import views  # noqa
